+++ Fovea.Renderer/Materials/Metal.cs
@@ -2,7 +2,6 @@ using System;
 using Fovea.Renderer.Core;
 using Fovea.Renderer.Image;
 using Fovea.Renderer.Sampling;
-using Fovea.Renderer.VectorMath;
 
 namespace Fovea.Renderer.Materials
 {
@@ -28,7 +27,7 @@ namespace Fovea.Renderer.Materials
             scatterResult.Attenuation = _albedo.Value(hitRecord.TextureU, hitRecord.TextureV, hitRecord.HitPoint);
 
             var reflected = Vector3.Reflect(Vector3.Normalize(rayIn.Direction), hitRecord.Normal);
-            var dir = reflected + Sampler.Instance.RandomOnUnitSphere() * (float)_fuzzy;
+            var dir = reflected + Sampler.Instance.RandomOnUnitSphere() * _fuzzy;
             scatterResult.SpecularRay =
                 new Ray(hitRecord.HitPoint, dir, rayIn.Time);
             return true;
