+++ Fovea.Renderer/Primitives/MovingSphere.cs
@@ -26,7 +26,7 @@ namespace Fovea.Renderer.Primitives
 
         public bool Hit(in Ray ray, in Interval rayInterval, ref HitRecord hitRecord)
         {
-            var center = CenterAtTime((float)ray.Time);
+            var center = CenterAtTime(ray.Time);
             var root = 0.0f;
             if (!Sphere.IntersectSphere(ray, center, _radius, rayInterval, ref root))
                 return false;
